[dataio] Check if the input format is supported with the "image" module
For ImageFeed: use image::isSupported and image::isVideoExtension to check
that the input is supported by OpenImageIO but is not a video.

For VideoFeed: add a VideoFeed::isSupported method to ensure that we do
not try to open unsupported videos with OpenCV. The list of supported
video formats is provided by OpenImageIO, which is based on ffmpeg like
OpenCV.

Both ImageFeed and VideoFeed have their own implementation of isSupported
to ensure they can check whether they support a given input on their own.